Split out test, disable it for Mono
