+++ src/tests/JIT/Directed/callconv/ThisCall/EmptyThisCallTest.csproj
@@ -4,7 +4,7 @@
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="ThisCallTest.cs" />
+    <Compile Include="$(MSBuildProjectName).cs" />
   </ItemGroup>
   <ItemGroup>
     <CMakeProjectReference Include="../CMakeLists.txt" />
