+++ src/tests/JIT/Directed/callconv/ThisCall/ThisCallTest.cs
@@ -74,9 +74,6 @@ unsafe class ThisCallNative
     public static extern E GetEFromManaged(C* c);
     [DllImport(nameof(ThisCallNative))]
     public static extern CLong GetWidthAsLongFromManaged(C* c);
-
-    [DllImport(nameof(ThisCallNative), CallingConvention = CallingConvention.ThisCall, EntryPoint = "GetWidthAsLongFromManaged")]
-    public static extern int ThisCallWithEmptySignature();
 }
 
 unsafe class ThisCallTest
@@ -103,7 +100,6 @@ unsafe class ThisCallTest
             Test4ByteNonHFAUnmanagedCallersOnly();
             TestEnumUnmanagedCallersOnly();
             TestCLongUnmanagedCallersOnly();
-            TestEmpty();
         }
         catch (System.Exception ex)
         {
@@ -344,39 +340,5 @@ unsafe class ThisCallTest
     private static CLong GetWidthAsLong(ThisCallNative.C* c)
     {
         return new CLong((nint)c->width);
-    }
-
-    private static void TestEmpty()
-    {
-        try
-        {
-            delegate* unmanaged[Thiscall]<ThisCallNative.C*, CLong> fn = &GetWidthAsLong;
-            CalliEmptyThisCall((delegate* unmanaged[Thiscall]<int>)fn);
-            throw new Exception("FAIL: thiscall fptr with no args should have failed");
-        }
-        catch (InvalidProgramException)
-        {
-            Console.WriteLine("thiscall fptr with no args failed as expected");
-        }
-
-        try
-        {
-            PinvokeEmptyThisCall();
-            throw new Exception("FAIL: pinvoke thiscall with no args should have failed");
-        }
-        catch (InvalidProgramException)
-        {
-            Console.WriteLine("thiscall pinvoke with no args failed as expected");
-        }
-    }
-
-    private static int CalliEmptyThisCall(delegate* unmanaged[Thiscall]<int> fn)
-    {
-        return fn();
-    }
-
-    private static void PinvokeEmptyThisCall()
-    {
-        ThisCallNative.ThisCallWithEmptySignature();
     }
 }
